Settings classes moved form files containing direct initialisations
