+++ src/matchbox/server/postgresql/__init__.py
@@ -1,8 +1,8 @@
 """PostgreSQL adapter for Matchbox server."""
 
+from matchbox.server.postgresql._settings import MatchboxPostgresSettings
 from matchbox.server.postgresql.adapter import (
     MatchboxPostgres,
-    MatchboxPostgresSettings,
 )
 
 __all__ = ["MatchboxPostgres", "MatchboxPostgresSettings"]
