+++ src/matchbox/server/postgresql/_settings.py
@@ -1,0 +1,27 @@
+from pydantic import BaseModel, Field
+
+from matchbox.server.base import MatchboxBackends, MatchboxSettings
+
+
+class MatchboxPostgresCoreSettings(BaseModel):
+    """PostgreSQL-specific settings for Matchbox."""
+
+    host: str
+    port: int
+    user: str
+    password: str
+    database: str
+    db_schema: str
+
+
+class MatchboxPostgresSettings(MatchboxSettings):
+    """Settings for the Matchbox PostgreSQL backend.
+
+    Inherits the core settings and adds the PostgreSQL-specific settings.
+    """
+
+    backend_type: MatchboxBackends = MatchboxBackends.POSTGRES
+
+    postgres: MatchboxPostgresCoreSettings = Field(
+        default_factory=MatchboxPostgresCoreSettings
+    )
