+++ src/matchbox/server/postgresql/db.py
@@ -1,34 +1,9 @@
 """Matchbox PostgreSQL database connection."""
 
-from pydantic import BaseModel, Field
 from sqlalchemy import Engine, MetaData, create_engine, text
 from sqlalchemy.orm import declarative_base, sessionmaker
 
-from matchbox.server.base import MatchboxBackends, MatchboxSettings
-
-
-class MatchboxPostgresCoreSettings(BaseModel):
-    """PostgreSQL-specific settings for Matchbox."""
-
-    host: str
-    port: int
-    user: str
-    password: str
-    database: str
-    db_schema: str
-
-
-class MatchboxPostgresSettings(MatchboxSettings):
-    """Settings for the Matchbox PostgreSQL backend.
-
-    Inherits the core settings and adds the PostgreSQL-specific settings.
-    """
-
-    backend_type: MatchboxBackends = MatchboxBackends.POSTGRES
-
-    postgres: MatchboxPostgresCoreSettings = Field(
-        default_factory=MatchboxPostgresCoreSettings
-    )
+from matchbox.server.postgresql._settings import MatchboxPostgresSettings
 
 
 class MatchboxDatabase:
